instagram/requirements.txt
selenium==4.15.2
playwright==1.40.0
requests==2.31.0
python-telegram-bot==20.7
schedule==1.2.0
python-dotenv==1.0.0
webdriver-manager==4.0.1 